requirements.txt
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
greenlet==3.0.3
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
schedule==1.2.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wait-for-it==2.2.2
wsproto==1.2.0
